moved existingUser to UserGeneratorController for clean architecture purposes
